routerify everything
