+++ api/src/main.py
@@ -54,11 +54,20 @@ from slowapi.errors import RateLimitExceeded
 from slowapi.middleware import SlowAPIMiddleware
 from slowapi.util import get_remote_address
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import Session
 from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from . import agent, crud, schemas
-from .db import SessionLocal, engine, scaffold_db
+from .db import engine, scaffold_db
+from src.dependencies import db
+from src.routers import (
+    apps,
+    users,
+    sessions,
+    messages,
+    metamessages,
+    collections,
+    documents,
+)
 
 # Otel Setup
 
@@ -220,15 +229,6 @@ app.add_middleware(SlowAPIMiddleware)
 add_pagination(app)
 
 
-async def get_db():
-    """FastAPI Dependency Generator for Database"""
-    db: AsyncSession = SessionLocal()
-    try:
-        yield db
-    finally:
-        await db.close()
-
-
 @app.exception_handler(StarletteHTTPException)
 async def http_exception_handler(request, exc):
     current_span = trace.get_current_span()
@@ -245,1033 +245,10 @@ async def http_exception_handler(request, exc):
     )
 
 
-########################################################
-# App Routes
-########################################################
-@app.get("/apps/{app_id}", response_model=schemas.App)
-async def get_app(
-    request: Request,
-    app_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get an App by ID
-
-    Args:
-        app_id (uuid.UUID): The ID of the app
-
-    Returns:
-        schemas.App: App object
-
-    """
-    app = await crud.get_app(db, app_id=app_id)
-    if app is None:
-        raise HTTPException(status_code=404, detail="App not found")
-    return app
-
-
-@app.get("/apps/name/{name}", response_model=schemas.App)
-async def get_app_by_name(
-    request: Request,
-    name: str,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get an App by Name
-
-    Args:
-        app_name (str): The name of the app
-
-    Returns:
-        schemas.App: App object
-
-    """
-    app = await crud.get_app_by_name(db, name=name)
-    if app is None:
-        raise HTTPException(status_code=404, detail="App not found")
-    return app
-
-
-@app.post("/apps", response_model=schemas.App)
-async def create_app(
-    request: Request,
-    app: schemas.AppCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Create an App
-
-    Args:
-        app (schemas.AppCreate): The App object containing any metadata
-
-    Returns:
-        schemas.App: Created App object
-
-    """
-
-    return await crud.create_app(db, app=app)
-
-
-@app.get("/apps/get_or_create/{name}", response_model=schemas.App)
-async def get_or_create_app(
-    request: Request,
-    name: str,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get or Create an App
-
-    Args:
-        app_name (str): The name of the app
-
-    Returns:
-        schemas.App: App object
-
-    """
-    app = await crud.get_app_by_name(db, name=name)
-    if app is None:
-        app = await crud.create_app(db, app=schemas.AppCreate(name=name))
-    return app
-
-
-@app.put("/apps/{app_id}", response_model=schemas.App)
-async def update_app(
-    request: Request,
-    app_id: uuid.UUID,
-    app: schemas.AppUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Update an App
-
-    Args:
-        app_id (uuid.UUID): The ID of the app to update
-        app (schemas.AppUpdate): The App object containing any new metadata
-
-    Returns:
-        schemas.App: The App object of the updated App
-
-    """
-    honcho_app = await crud.update_app(db, app_id=app_id, app=app)
-    if honcho_app is None:
-        raise HTTPException(status_code=404, detail="App not found")
-    return honcho_app
-
-
-########################################################
-# User Routes
-########################################################
-
-
-@app.post("/apps/{app_id}/users", response_model=schemas.User)
-async def create_user(
-    request: Request,
-    app_id: uuid.UUID,
-    user: schemas.UserCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Create a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user (schemas.UserCreate): The User object containing any metadata
-
-    Returns:
-        schemas.User: Created User object
-
-    """
-    return await crud.create_user(db, app_id=app_id, user=user)
-
-
-@app.get("/apps/{app_id}/users", response_model=Page[schemas.User])
-async def get_users(
-    request: Request,
-    app_id: uuid.UUID,
-    reverse: bool = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get All Users for an App
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client
-        application using honcho
-
-    Returns:
-        list[schemas.User]: List of User objects
-
-    """
-    data = None
-    if filter is not None:
-        data = json.loads(filter)
-
-    return await paginate(
-        db, await crud.get_users(db, app_id=app_id, reverse=reverse, filter=data)
-    )
-
-
-@app.get("/apps/{app_id}/users/{name}", response_model=schemas.User)
-async def get_user_by_name(
-    request: Request,
-    app_id: uuid.UUID,
-    name: str,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-
-    Returns:
-        schemas.User: User object
-
-    """
-    return await crud.get_user_by_name(db, app_id=app_id, name=name)
-
-
-@app.get("/apps/{app_id}/users/get_or_create/{name}", response_model=schemas.User)
-async def get_or_create_user(
-    request: Request, app_id: uuid.UUID, name: str, db: AsyncSession = Depends(get_db)
-):
-    """Get or Create a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-
-    Returns:
-        schemas.User: User object
-
-    """
-    user = await crud.get_user_by_name(db, app_id=app_id, name=name)
-    if user is None:
-        user = await crud.create_user(
-            db, app_id=app_id, user=schemas.UserCreate(name=name)
-        )
-    return user
-
-
-@app.put("/apps/{app_id}/users/{user_id}", response_model=schemas.User)
-async def update_user(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    user: schemas.UserUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Update a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        user (schemas.UserCreate): The User object containing any metadata
-
-    Returns:
-        schemas.User: Updated User object
-
-    """
-    return await crud.update_user(db, app_id=app_id, user_id=user_id, user=user)
-
-
-########################################################
-# Session Routes
-########################################################
-
-
-@router.get("/sessions", response_model=Page[schemas.Session])
-async def get_sessions(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    location_id: Optional[str] = None,
-    is_active: Optional[bool] = False,
-    reverse: Optional[bool] = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get All Sessions for a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-        location_id (str, optional): Optional Location ID representing the location of a
-        session
-
-    Returns:
-        list[schemas.Session]: List of Session objects
-
-    """
-
-    data = None
-    if filter is not None:
-        data = json.loads(filter)
-
-    return await paginate(
-        db,
-        await crud.get_sessions(
-            db,
-            app_id=app_id,
-            user_id=user_id,
-            location_id=location_id,
-            reverse=reverse,
-            is_active=is_active,
-            filter=data,
-        ),
-    )
-
-
-@router.post("/sessions", response_model=schemas.Session)
-async def create_session(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session: schemas.SessionCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Create a Session for a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client
-        application using honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-        session (schemas.SessionCreate): The Session object containing any
-        metadata and a location ID
-
-    Returns:
-        schemas.Session: The Session object of the new Session
-
-    """
-    value = await crud.create_session(
-        db, app_id=app_id, user_id=user_id, session=session
-    )
-    return value
-
-
-@router.put("/sessions/{session_id}", response_model=schemas.Session)
-async def update_session(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    session: schemas.SessionUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Update the metadata of a Session
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-        session_id (uuid.UUID): The ID of the Session to update
-        session (schemas.SessionUpdate): The Session object containing any new metadata
-
-    Returns:
-        schemas.Session: The Session object of the updated Session
-
-    """
-    if session.metadata is None:
-        raise HTTPException(status_code=400, detail="Session metadata cannot be empty")
-    try:
-        return await crud.update_session(
-            db, app_id=app_id, user_id=user_id, session_id=session_id, session=session
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-@router.delete("/sessions/{session_id}")
-async def delete_session(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    """Delete a session by marking it as inactive
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-        session_id (uuid.UUID): The ID of the Session to delete
-
-    Returns:
-        dict: A message indicating that the session was deleted
-
-    Raises:
-        HTTPException: If the session is not found
-
-    """
-    response = await crud.delete_session(
-        db, app_id=app_id, user_id=user_id, session_id=session_id
-    )
-    if response:
-        return {"message": "Session deleted successfully"}
-    else:
-        raise HTTPException(status_code=404, detail="Session not found")
-
-
-@router.get("/sessions/{session_id}", response_model=schemas.Session)
-async def get_session(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get a specific session for a user by ID
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-        session_id (uuid.UUID): The ID of the Session to retrieve
-
-    Returns:
-        schemas.Session: The Session object of the requested Session
-
-    Raises:
-        HTTPException: If the session is not found
-    """
-    honcho_session = await crud.get_session(
-        db, app_id=app_id, session_id=session_id, user_id=user_id
-    )
-    if honcho_session is None:
-        raise HTTPException(status_code=404, detail="Session not found")
-    return honcho_session
-
-
-########################################################
-# Message Routes
-########################################################
-
-
-@router.post("/sessions/{session_id}/messages", response_model=schemas.Message)
-async def create_message_for_session(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    message: schemas.MessageCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Adds a message to a session
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        session_id (int): The ID of the Session to add the message to
-        message (schemas.MessageCreate): The Message object to add containing the
-        message content and type
-
-    Returns:
-        schemas.Message: The Message object of the added message
-
-    Raises:
-        HTTPException: If the session is not found
-
-    """
-    try:
-        return await crud.create_message(
-            db, message=message, app_id=app_id, user_id=user_id, session_id=session_id
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-@router.get("/sessions/{session_id}/messages", response_model=Page[schemas.Message])
-async def get_messages(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    reverse: Optional[bool] = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get all messages for a session
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        session_id (int): The ID of the Session to retrieve
-        reverse (bool): Whether to reverse the order of the messages
-
-    Returns:
-        list[schemas.Message]: List of Message objects
-
-    Raises:
-        HTTPException: If the session is not found
-
-    """
-    try:
-        data = None
-        if filter is not None:
-            data = json.loads(filter)
-        return await paginate(
-            db,
-            await crud.get_messages(
-                db,
-                app_id=app_id,
-                user_id=user_id,
-                session_id=session_id,
-                filter=data,
-                reverse=reverse,
-            ),
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-@router.get(
-    "sessions/{session_id}/messages/{message_id}", response_model=schemas.Message
-)
-async def get_message(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    message_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    """ """
-    honcho_message = await crud.get_message(
-        db, app_id=app_id, session_id=session_id, user_id=user_id, message_id=message_id
-    )
-    if honcho_message is None:
-        raise HTTPException(status_code=404, detail="Session not found")
-    return honcho_message
-
-
-@router.put(
-    "sessions/{session_id}/messages/{message_id}", response_model=schemas.Message
-)
-async def update_message(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    message_id: uuid.UUID,
-    message: schemas.MessageUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Update's the metadata of a message"""
-    if message.metadata is None:
-        raise HTTPException(status_code=400, detail="Message metadata cannot be empty")
-    try:
-        return await crud.update_message(
-            db,
-            message=message,
-            app_id=app_id,
-            user_id=user_id,
-            session_id=session_id,
-            message_id=message_id,
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-########################################################
-# metamessage routes
-########################################################
-
-
-@router.post("/sessions/{session_id}/metamessages", response_model=schemas.Metamessage)
-async def create_metamessage(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    metamessage: schemas.MetamessageCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Adds a message to a session
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        session_id (int): The ID of the Session to add the message to
-        message (schemas.MessageCreate): The Message object to add containing the
-        message content and type
-
-    Returns:
-        schemas.Message: The Message object of the added message
-
-    Raises:
-        HTTPException: If the session is not found
-
-    """
-    try:
-        return await crud.create_metamessage(
-            db,
-            metamessage=metamessage,
-            app_id=app_id,
-            user_id=user_id,
-            session_id=session_id,
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-@router.get(
-    "/sessions/{session_id}/metamessages", response_model=Page[schemas.Metamessage]
-)
-async def get_metamessages(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    message_id: Optional[uuid.UUID] = None,
-    metamessage_type: Optional[str] = None,
-    reverse: Optional[bool] = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get all messages for a session
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        session_id (int): The ID of the Session to retrieve
-        reverse (bool): Whether to reverse the order of the metamessages
-
-    Returns:
-        list[schemas.Message]: List of Message objects
-
-    Raises:
-        HTTPException: If the session is not found
-
-    """
-    try:
-        data = None
-        if filter is not None:
-            data = json.loads(filter)
-        return await paginate(
-            db,
-            await crud.get_metamessages(
-                db,
-                app_id=app_id,
-                user_id=user_id,
-                session_id=session_id,
-                message_id=message_id,
-                metamessage_type=metamessage_type,
-                filter=data,
-                reverse=reverse,
-            ),
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-@router.get(
-    "/sessions/{session_id}/metamessages/{metamessage_id}",
-    response_model=schemas.Metamessage,
-)
-async def get_metamessage(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    message_id: uuid.UUID,
-    metamessage_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get a specific Metamessage by ID
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client application using
-        honcho
-        user_id (str): The User ID representing the user, managed by the user
-        session_id (int): The ID of the Session to retrieve
-
-    Returns:
-        schemas.Session: The Session object of the requested Session
-
-    Raises:
-        HTTPException: If the session is not found
-    """
-    honcho_metamessage = await crud.get_metamessage(
-        db,
-        app_id=app_id,
-        session_id=session_id,
-        user_id=user_id,
-        message_id=message_id,
-        metamessage_id=metamessage_id,
-    )
-    if honcho_metamessage is None:
-        raise HTTPException(status_code=404, detail="Session not found")
-    return honcho_metamessage
-
-
-@router.put(
-    "sessions/{session_id}/metamessages/{metamessage_id}",
-    response_model=schemas.Metamessage,
-)
-async def update_metamessage(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    metamessage_id: uuid.UUID,
-    metamessage: schemas.MetamessageUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    """Update's the metadata of a metamessage"""
-    if metamessage.metadata is None:
-        raise HTTPException(
-            status_code=400, detail="Metamessage metadata cannot be empty"
-        )
-    try:
-        return await crud.update_metamessage(
-            db,
-            metamessage=metamessage,
-            app_id=app_id,
-            user_id=user_id,
-            session_id=session_id,
-            metamessage_id=metamessage_id,
-        )
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Session not found") from None
-
-
-########################################################
-# collection routes
-########################################################
-
-
-@router.get("/collections", response_model=Page[schemas.Collection])
-async def get_collections(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    reverse: Optional[bool] = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    """Get All Collections for a User
-
-    Args:
-        app_id (uuid.UUID): The ID of the app representing the client
-        application using honcho
-        user_id (uuid.UUID): The User ID representing the user, managed by the user
-
-    Returns:
-        list[schemas.Collection]: List of Collection objects
-
-    """
-    data = None
-    if filter is not None:
-        data = json.loads(filter)
-    return await paginate(
-        db,
-        await crud.get_collections(
-            db, app_id=app_id, user_id=user_id, filter=data, reverse=reverse
-        ),
-    )
-
-
-# @router.get("/collections/id/{collection_id}", response_model=schemas.Collection)
-# def get_collection_by_id(
-#     request: Request,
-#     app_id: uuid.UUID,
-#     user_id: uuid.UUID,
-#     collection_id: uuid.UUID,
-#     db: AsyncSession = Depends(get_db),
-# ) -> schemas.Collection:
-#     honcho_collection = crud.get_collection_by_id(
-#         db, app_id=app_id, user_id=user_id, collection_id=collection_id
-#     )
-#     if honcho_collection is None:
-#         raise HTTPException(
-#             status_code=404, detail="collection not found or does not belong to user"
-#         )
-#     return honcho_collection
-
-
-@router.get("/collections/{name}", response_model=schemas.Collection)
-async def get_collection_by_name(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    name: str,
-    db: AsyncSession = Depends(get_db),
-) -> schemas.Collection:
-    honcho_collection = await crud.get_collection_by_name(
-        db, app_id=app_id, user_id=user_id, name=name
-    )
-    if honcho_collection is None:
-        raise HTTPException(
-            status_code=404, detail="collection not found or does not belong to user"
-        )
-    return honcho_collection
-
-
-@router.post("/collections", response_model=schemas.Collection)
-async def create_collection(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection: schemas.CollectionCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    if collection.name == "honcho":
-        raise HTTPException(
-            status_code=406,
-            detail="error invalid collection configuration - honcho is a reserved name",
-        )
-    try:
-        return await crud.create_collection(
-            db, collection=collection, app_id=app_id, user_id=user_id
-        )
-    except ValueError:
-        raise HTTPException(
-            status_code=406,
-            detail="Error invalid collection configuration - name may already exist",
-        ) from None
-
-
-@router.put("/collections/{collection_id}", response_model=schemas.Collection)
-async def update_collection(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    collection: schemas.CollectionUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    if collection.name is None:
-        raise HTTPException(
-            status_code=400, detail="invalid request - name cannot be None"
-        )
-    if collection.name == "honcho":
-        raise HTTPException(
-            status_code=406,
-            detail="error invalid collection configuration - honcho is a reserved name",
-        )
-
-    try:
-        honcho_collection = await crud.update_collection(
-            db,
-            collection=collection,
-            app_id=app_id,
-            user_id=user_id,
-            collection_id=collection_id,
-        )
-    except ValueError:
-        raise HTTPException(
-            status_code=406,
-            detail="Error invalid collection configuration - name may already exist",
-        ) from None
-    return honcho_collection
-
-
-@router.delete("/collections/{collection_id}")
-async def delete_collection(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    response = await crud.delete_collection(
-        db, app_id=app_id, user_id=user_id, collection_id=collection_id
-    )
-    if response:
-        return {"message": "Collection deleted successfully"}
-    else:
-        raise HTTPException(
-            status_code=404, detail="collection not found or does not belong to user"
-        )
-
-
-########################################################
-# Document routes
-########################################################
-
-
-@router.get(
-    "/collections/{collection_id}/documents", response_model=Page[schemas.Document]
-)
-async def get_documents(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    reverse: Optional[bool] = False,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    try:
-        data = None
-        if filter is not None:
-            data = json.loads(filter)
-        return await paginate(
-            db,
-            await crud.get_documents(
-                db,
-                app_id=app_id,
-                user_id=user_id,
-                collection_id=collection_id,
-                filter=data,
-                reverse=reverse,
-            ),
-        )
-    except (
-        ValueError
-    ):  # TODO can probably remove this exception ok to return empty here
-        raise HTTPException(
-            status_code=404, detail="collection not found or does not belong to user"
-        ) from None
-
-
-router.get(
-    "/collections/{collection_id}/documents/{document_id}",
-    response_model=schemas.Document,
-)
-
-
-async def get_document(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    document_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    honcho_document = await crud.get_document(
-        db,
-        app_id=app_id,
-        user_id=user_id,
-        collection_id=collection_id,
-        document_id=document_id,
-    )
-    if honcho_document is None:
-        raise HTTPException(
-            status_code=404, detail="document not found or does not belong to user"
-        )
-    return honcho_document
-
-
-@router.get(
-    "/collections/{collection_id}/query", response_model=Sequence[schemas.Document]
-)
-async def query_documents(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    query: str,
-    top_k: int = 5,
-    filter: Optional[str] = None,
-    db: AsyncSession = Depends(get_db),
-):
-    if top_k is not None and top_k > 50:
-        top_k = 50  # TODO see if we need to paginate this
-    data = None
-    if filter is not None:
-        data = json.loads(filter)
-    return await crud.query_documents(
-        db=db,
-        app_id=app_id,
-        user_id=user_id,
-        collection_id=collection_id,
-        query=query,
-        filter=data,
-        top_k=top_k,
-    )
-
-
-@router.post("/collections/{collection_id}/documents", response_model=schemas.Document)
-async def create_document(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    document: schemas.DocumentCreate,
-    db: AsyncSession = Depends(get_db),
-):
-    try:
-        return await crud.create_document(
-            db,
-            document=document,
-            app_id=app_id,
-            user_id=user_id,
-            collection_id=collection_id,
-        )
-    except ValueError:
-        raise HTTPException(
-            status_code=404, detail="collection not found or does not belong to user"
-        ) from None
-
-
-@router.put(
-    "/collections/{collection_id}/documents/{document_id}",
-    response_model=schemas.Document,
-)
-async def update_document(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    document_id: uuid.UUID,
-    document: schemas.DocumentUpdate,
-    db: AsyncSession = Depends(get_db),
-):
-    if document.content is None and document.metadata is None:
-        raise HTTPException(
-            status_code=400, detail="content and metadata cannot both be None"
-        )
-    return await crud.update_document(
-        db,
-        document=document,
-        app_id=app_id,
-        user_id=user_id,
-        collection_id=collection_id,
-        document_id=document_id,
-    )
-
-
-@router.delete("/collections/{collection_id}/documents/{document_id}")
-async def delete_document(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    collection_id: uuid.UUID,
-    document_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    response = await crud.delete_document(
-        db,
-        app_id=app_id,
-        user_id=user_id,
-        collection_id=collection_id,
-        document_id=document_id,
-    )
-    if response:
-        return {"message": "Document deleted successfully"}
-    else:
-        raise HTTPException(
-            status_code=404, detail="document not found or does not belong to user"
-        )
-
-
-@router.get("/sessions/{session_id}/chat", response_model=schemas.AgentChat)
-async def get_chat(
-    request: Request,
-    app_id: uuid.UUID,
-    user_id: uuid.UUID,
-    session_id: uuid.UUID,
-    query: str,
-    db: AsyncSession = Depends(get_db),
-):
-    return await agent.chat(
-        app_id=app_id, user_id=user_id, session_id=session_id, query=query, db=db
-    )
-
-
-app.include_router(router)
+app.include_router(apps.router)
+app.include_router(users.router)
+app.include_router(sessions.router)
+app.include_router(messages.router)
+app.include_router(metamessages.router)
+app.include_router(collections.router)
+app.include_router(documents.router)
